+++ foodb_isolate_example/android/app/src/main/kotlin/cc/feedme/foodb_isolate_example/MainActivity.kt
@@ -1,0 +1,6 @@
+package cc.feedme.foodb_isolate_example
+
+import io.flutter.embedding.android.FlutterActivity
+
+class MainActivity: FlutterActivity() {
+}