+++ foodb_test/lib/src/test/get_test.dart
@@ -1,10 +1,12 @@
+import 'dart:io';
+
 import 'package:test/test.dart';
 import 'package:foodb/foodb.dart';
 import 'package:foodb_test/foodb_test.dart';
 
 void main() {
-  // final ctx = CouchdbTestContext();
-  final ctx = InMemoryTestContext();
+  final ctx = CouchdbTestContext();
+  // final ctx = InMemoryTestContext();
   getTest().forEach((t) {
     t(ctx);
   });
@@ -101,6 +103,32 @@ List<Function(FoodbTestContext)> getTest() {
                 element.docs.every((element) => element.error != null)),
             hasLength(1));
       });
-    }
+    },
+    (FoodbTestContext ctx) {
+      test('bulkGet() improve json decode', () async {
+        FoodbDebug.logLevel = LOG_LEVEL.debug;
+        final db = await ctx.db('bulk-get-large-json', persist: true);
+        if ((await db.info()).docCount == 0) {
+          var docs = <Doc<Map<String, dynamic>>>[];
+          for (var i = 0; i <= 500; ++i) {
+            docs.add(getLargeDoc('$i'));
+            if (i % 100 == 0) {
+              await db.bulkDocs(body: docs);
+              docs = [];
+            }
+          }
+        }
+        final allDoc = await db.allDocs(GetViewRequest(), (json) => json);
+
+        final result = await db.bulkGet(
+            body: BulkGetRequest(
+                docs: allDoc.rows
+                    .map((e) => BulkGetRequestDoc(
+                        id: e.id, rev: Rev.fromString(e.value['rev'])))
+                    .toList()),
+            fromJsonT: (json) => json);
+        expect(result, isNotNull);
+      }, timeout: Timeout(Duration(minutes: 5)));
+    },
   ];
 }